+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     * Fixed `ignore_empty: true` not removing empty values in blueprint filtering
     * Fixed `{{ false|string }}` twig to return '0' instead of ''
     * Fixed `Data::filter()` removing empty fields (such as empty list) by default
+    * Fixed twig `url()` failing if stream has extra slash in it (e.g. `user:///data`)
     * Grav 1.7: Fixed `Flex Pages` unserialize issues if Flex-Objects Plugin has not been installed
     * Grav 1.7: Require Flex-Objects Plugin to edit Flex Accounts